refactor: CourseItem type 로직 분리
